refactor : 사소한 오류 수정
