+++ src/main/java/com/hanwul/kbscbackend/domain/mission/success/Success.java
@@ -4,6 +4,7 @@ import com.hanwul.kbscbackend.common.BaseEntity;
 import com.hanwul.kbscbackend.domain.account.Account;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
@@ -12,6 +13,7 @@ import javax.persistence.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+@Getter
 @Entity
 public class Success extends BaseEntity {
