+++ src/main/java/com/hanwul/kbscbackend/domain/question/QuestionController.java
@@ -1,5 +1,6 @@
 package com.hanwul.kbscbackend.domain.question;
 
+import com.hanwul.kbscbackend.domain.questionanswer.question.QuestionService;
 import com.hanwul.kbscbackend.dto.BasicResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
